test(api): split the RI exchange idempotency tests under the file-size limit (#1642)
ri_exchange_idempotency_test.go had grown to 684 lines, over the 500-line
limit in CLAUDE.md. The sibling handler test files in this package are far
worse (2013 and 2225 lines), but pre-existing debt does not license adding
more, so the new file is split along the seams it already had:

  ri_exchange_idempotency_test.go         the exchangeClaimLedger harness,
                                          the handler double-spend tests and
                                          the claim-outcome contract
  ri_exchange_idempotency_key_test.go     what the fingerprint is made of
  ri_exchange_idempotency_guards_test.go  the input checks that must refuse
                                          a request before it can claim

A pure relocation: comparing the sorted lines of the old file against the
three new ones shows only additions, all of them the new file headers. No
test body, assertion or expectation changed, and the set of declared
functions is identical.
